requirements.txt
amqp==5.0.6
app-store-scraper==0.3.5
asgiref==3.4.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
crispy-tailwind==0.5.0
cryptography==3.4.7
Django==3.2.5
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
flower==1.0.0
google-play-scraper==1.0.0
humanize==3.10.0
idna==2.10
itunes-app-scraper-dmi==0.9.4
kombu==5.1.0
lxml==4.6.3
play-scraper==0.6.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycparser==2.20
pyOpenSSL==20.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.23.0
requests-futures==1.0.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
